pyproject.toml
[project]
name = "sync"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pathvalidate>=3.3.1",
    "python-dotenv>=1.1.1",
]
